+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add validating webhook.
+  
+### Changed
+
+- Update manifests with upstream `v5.0.1`.
+- Use `kustomize` for templating upstream manifests without manual editing.
+
 ## [0.1.0] - 2022-04-25
 
 - Initial implementation.